Refactor cache infrastructure
