+++ cms/cache/permissions.py
@@ -16,12 +16,14 @@ def get_cache_key(user, key):
     return "%s:permission:%s:%s" % (
         get_cms_setting('CACHE_PREFIX'), username, key)
 
-def get_cache_version_key():
+
+def get_cache_permission_version_key():
     return "%s:permission:version" % (get_cms_setting('CACHE_PREFIX'),)
 
-def get_cache_version():
+
+def get_cache_permission_version():
     from django.core.cache import cache
-    version = cache.get(get_cache_version_key())
+    version = cache.get(get_cache_permission_version_key())
     if version is None:
         version = 1
     return version
@@ -32,7 +34,7 @@ def get_permission_cache(user, key):
     Helper for reading values from cache
     """
     from django.core.cache import cache
-    return cache.get(get_cache_key(user, key), version=get_cache_version())
+    return cache.get(get_cache_key(user, key), version=get_cache_permission_version())
 
 
 def set_permission_cache(user, key, value):
@@ -44,8 +46,8 @@ def set_permission_cache(user, key, value):
     # store this key, so we can clean it when required
     cache_key = get_cache_key(user, key)
     cache.set(cache_key, value,
-            get_cms_setting('CACHE_DURATIONS')['permissions'],
-            version=get_cache_version())
+              get_cms_setting('CACHE_DURATIONS')['permissions'],
+              version=get_cache_permission_version())
 
 
 def clear_user_permission_cache(user):
@@ -54,14 +56,14 @@ def clear_user_permission_cache(user):
     """
     from django.core.cache import cache
     for key in PERMISSION_KEYS:
-        cache.delete(get_cache_key(user, key), version=get_cache_version())
+        cache.delete(get_cache_key(user, key), version=get_cache_permission_version())
 
 
 def clear_permission_cache():
     from django.core.cache import cache
-    version = get_cache_version()
+    version = get_cache_permission_version()
     if version > 1:
-        cache.incr(get_cache_version_key())
+        cache.incr(get_cache_permission_version_key())
     else:
-        cache.set(get_cache_version_key(), 2,
-                get_cms_setting('CACHE_DURATIONS')['permissions'])
+        cache.set(get_cache_permission_version_key(), 2,
+                  get_cms_setting('CACHE_DURATIONS')['permissions'])
