+++ cms/forms/fields.py
@@ -1,12 +1,14 @@
 # -*- coding: utf-8 -*-
 import six
+
 from django.utils.translation import ugettext_lazy as _
 from django import forms
 from django.contrib.admin.widgets import RelatedFieldWidgetWrapper
 from django.forms.fields import EMPTY_VALUES
+
+from cms.cache.choices import get_site_choices, get_page_choices
 from cms.models.pagemodel import Page
 from cms.forms.widgets import PageSelectWidget, PageSmartLinkWidget
-from cms.forms.utils import get_site_choices, get_page_choices
 
 
 class SuperLazyIterator(object):
