+++ cms/tests/multilingual.py
@@ -9,8 +9,8 @@ from django.http import Http404, HttpResponseRedirect
 from django.test.utils import override_settings
 
 from cms.api import create_page, create_title, publish_page, add_plugin
+from cms.cache.choices import update_site_and_page_choices
 from cms.exceptions import LanguageError
-from cms.forms.utils import update_site_and_page_choices
 from cms.menu import CMSMenu
 from cms.models import Title, EmptyTitle
 from cms.test_utils.testcases import (CMSTestCase,
